+++ server/config/paymentconfig.js
@@ -1,0 +1,4 @@
+exports.PRODUCT_TYPES = {
+  "3ZRLN4LJVSRVY": "Landing Page",
+  "M3CSSZ43CE75J": "Portfolio Page",
+};
